+++ src/main/java/br/com/fiap/bookorganizer/models/Autor.java
@@ -10,12 +10,12 @@ public class Autor {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
+    private Long id;
     private String nome;
 
     protected Autor() {}
 
-    public Autor(Integer id, String nome) {
+    public Autor(Long id, String nome) {
         this.id = id;
         this.nome = nome;
     }
@@ -25,11 +25,11 @@ public class Autor {
         return "Autor [id=" + id + ", nome=" + nome + "]";
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
