+++ src/main/java/br/com/fiap/bookorganizer/models/Categoria.java
@@ -10,12 +10,12 @@ public class Categoria {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
+    private Long id;
     private String nome;
 
     protected Categoria() {}
 
-    public Categoria(Integer id, String nome) {
+    public Categoria(Long id, String nome) {
         this.id = id;
         this.nome = nome;
     }
@@ -25,10 +25,10 @@ public class Categoria {
         return "Categoria [id=" + id + ", nome=" + nome + "]";
     }
     
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
     public String getNome() {
